requirements.txt
streamlit==1.34.0
openai==1.30.3
psycopg2-binary==2.9.9
mysql-connector-python==8.3.0
python-dotenv==1.0.1
